add model for all entities
